split out DisallowWordsSystem into its own file
